hackhaton2/phase2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "python-dotenv>=0.21.0",
    "pytest>=7.4.0",
    "httpx>=0.27.0",
]
